// src/database/types/domain.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * question entity domain
 */
export interface Domain {
  /**
   * Domain ID
   */
  id?: number;
  /**
   * The name for the domain
   */
  name: string;
  /**
   * Description of Domain
   */
  description: string;
  /**
   * logical identifier for domain
   */
  identifier?: string;
  /**
   * enable users to add entries
   */
  allowUserEntries?: boolean;
  /**
   * simple toggle to allow users to see results
   */
  showResults?: boolean;
  /**
   * users can score their own entries
   */
  scoreOwnEntry?: boolean;
  /**
   * stop users from submitting scores
   */
  disableScoring?: boolean;
  /**
   * optional theme id to use for rendering the app
   */
  theme?: number;
  /**
   * Ranks that will be scored against each entity
   */
  ranks: {
    /**
     * Rank Name
     */
    name: string;
    /**
     * Minimum selectable value
     */
    min: number;
    /**
     * Maximum selectable value
     */
    max: number;
    [k: string]: unknown;
  }[];
}